+++ src/python/pants/backend/python/util_rules/pex_requirements.py
@@ -16,17 +16,21 @@ from pants.backend.python.util_rules.lockfile_metadata import (
     PythonLockfileMetadata,
     PythonLockfileMetadataV2,
 )
+from pants.core.goals.generate_lockfiles import GenerateToolLockfileSentinel
 from pants.core.util_rules.lockfile_metadata import InvalidLockfileError, LockfileMetadataValidation
+from pants.engine.engine_aware import EngineAwareParameter
 from pants.engine.fs import (
     CreateDigest,
     Digest,
     DigestContents,
+    DigestEntries,
     FileContent,
     FileEntry,
     GlobMatchErrorBehavior,
     PathGlobs,
 )
-from pants.engine.rules import Get, collect_rules, rule
+from pants.engine.rules import Get, MultiGet, collect_rules, rule
+from pants.engine.unions import UnionMembership
 from pants.util.docutil import bin_name, doc_url
 from pants.util.meta import frozen_after_init
 from pants.util.ordered_set import FrozenOrderedSet
@@ -270,6 +274,12 @@ class PexRequirements:
         return bool(self.req_strings)
 
 
+# NB: This is defined here so that our rule for ResolvePexConfigRequest -> ResolvePexConfig
+# can avoid an import cycle. We re-export this out of goals/lockfile.py.
+class GeneratePythonToolLockfileSentinel(GenerateToolLockfileSentinel):
+    pass
+
+
 @dataclass(frozen=True)
 class ResolvePexConfig:
     """Configuration from `[python]` that impacts how the resolve is created."""
@@ -285,10 +295,62 @@ class ResolvePexConfig:
 
 
 @dataclass(frozen=True)
-class ResolvePexConfigRequest:
+class ResolvePexConfigRequest(EngineAwareParameter):
     """Find all configuration from `[python]` that impacts how the resolve is created."""
 
     resolve_name: str
+
+    def debug_hint(self) -> str:
+        return self.resolve_name
+
+
+@rule
+async def determine_resolve_pex_config(
+    request: ResolvePexConfigRequest, python_setup: PythonSetup, union_membership: UnionMembership
+) -> ResolvePexConfig:
+    all_python_tool_resolve_names = tuple(
+        sentinel.resolve_name
+        for sentinel in union_membership.get(GenerateToolLockfileSentinel)
+        if issubclass(sentinel, GeneratePythonToolLockfileSentinel)
+    )
+
+    constraints_file: tuple[Digest, FileEntry] | None = None
+    _constraints_file_path = python_setup.resolves_to_constraints_file(
+        all_python_tool_resolve_names
+    ).get(request.resolve_name)
+    if _constraints_file_path:
+        _constraints_origin = softwrap(
+            f"""
+            the option `[python].resolves_to_constraints_file` for the resolve
+            '{request.resolve_name}'
+            """
+        )
+        _constraints_path_globs = PathGlobs(
+            [_constraints_file_path] if _constraints_file_path else [],
+            glob_match_error_behavior=GlobMatchErrorBehavior.error,
+            description_of_origin=_constraints_origin,
+        )
+        _constraints_digest, _constraints_digest_entries = await MultiGet(
+            Get(Digest, PathGlobs, _constraints_path_globs),
+            Get(DigestEntries, PathGlobs, _constraints_path_globs),
+        )
+
+        if len(_constraints_digest_entries) != 1:
+            raise ValueError(
+                softwrap(
+                    f"""
+                    Expected only one file from {_constraints_origin}, but matched:
+                    {_constraints_digest_entries}
+
+                    Did you use a glob like `*`?
+                    """
+                )
+            )
+        _constraints_file_entry = next(iter(_constraints_digest_entries))
+        assert isinstance(_constraints_file_entry, FileEntry)
+        constraints_file = (_constraints_digest, _constraints_file_entry)
+
+    return ResolvePexConfig(constraints_file=constraints_file)
 
 
 def should_validate_metadata(
